Introduce global and per-tenant flags to control rule evaluation concurrency
This change introduces a global boolean flag named `-ruler.concurrent-rule-evaluation` to control if Mimir runs independent rules concurrently.
In addition to this flag, a per tenant configuration option of `-ruler.max-concurrent-rule-evaluations` is also introduced to control the amount of concurrency we can have per tenant.

By default, the new feature is disabled globally and it can also be disabbled per tenant by using a value of `0` as part of `-ruler.max-concurrent-rule-evaluations`.

Signed-off-by: gotjosh <josue.abreu@gmail.com>